+++ groove-security/spring-security/src/main/java/org/bogus/groove/config/authentication/CustomDaoAuthenticationProvider.java
@@ -1,7 +1,8 @@
-package org.bogus.groove.config;
+package org.bogus.groove.config.authentication;
 
 import org.bogus.groove.common.ErrorType;
 import org.bogus.groove.common.UnauthorizedException;
+import org.bogus.groove.config.CustomUserDetails;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
 import org.springframework.security.core.Authentication;
