+++ groove-security/spring-security/src/main/java/org/bogus/groove/config/authorization/JwtAuthorizationFilter.java
@@ -1,4 +1,4 @@
-package org.bogus.groove.config;
+package org.bogus.groove.config.authorization;
 
 import java.io.IOException;
 import javax.servlet.FilterChain;
@@ -6,6 +6,7 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
+import org.bogus.groove.config.CustomUserDetails;
 import org.bogus.groove.domain.user.UserInfoFinder;
 import org.bogus.groove.domain.user.token.TokenValidator;
 import org.springframework.http.HttpHeaders;
